// interview-bot-app/types/mock-interview-message-response.ts
import { z } from "zod";

/**
 * Type definition for message generated by the mock interview bot.
 */
export const MockInterviewMessageResponseSchema = z.object({
  reasoning: z.string(),
  message: z.string(),
});

export type MockInterviewMessageResponse = z.infer<typeof MockInterviewMessageResponseSchema>;